+++ src/dkplsrda.jl
@@ -86,7 +86,7 @@ function predict(object::Dkplsrda, X; nlv = nothing)
     a = size(object.fm.fm.T, 2)
     isnothing(nlv) ? nlv = a : nlv = (max(minimum(nlv), 0):min(maximum(nlv), a))
     le_nlv = length(nlv)
-    pred = list(le_nlv, Union{Matrix{Int64}, Matrix{Float64}, Matrix{String}})
+    pred = list(le_nlv, Union{Matrix{Int}, Matrix{Float64}, Matrix{String}})
     posterior = list(le_nlv, Matrix{Float64})
     @inbounds for i = 1:le_nlv
         zp = predict(object.fm, X; nlv = nlv[i]).pred
