+++ src/lwplsr_s.jl
@@ -113,7 +113,7 @@ function lwplsr_s(X, Y; reduc = :pls,
     n = nro(X)
     s = 1:n
     if psamp < 1
-        m = Int64(round(psamp * n))
+        m = Int(round(psamp * n))
         if samp == :sys
             s = sampsys(rowsum(Y), m).train
         elseif samp == :rand
