+++ src/occsd.jl
@@ -150,7 +150,7 @@ function predict(object::Occsd, X)
     p_val = pval(object.e_cdf, d)
     d = DataFrame(d = d, dstand = d / object.cutoff, 
         pval = p_val, gh = d2 / nlv)
-    pred = reshape(Int64.(d.dstand .> 1), m, 1)
+    pred = reshape(Int.(d.dstand .> 1), m, 1)
     (pred = pred, d)
 end
