+++ src/plslda.jl
@@ -109,7 +109,7 @@ function predict(object::Plslda, X; nlv = nothing)
     a = size(object.fm.fm_pls.T, 2)
     isnothing(nlv) ? nlv = a : nlv = (max(minimum(nlv), 0):min(maximum(nlv), a))
     le_nlv = length(nlv)
-    pred = list(le_nlv, Union{Matrix{Int64}, Matrix{Float64}, Matrix{String}})
+    pred = list(le_nlv, Union{Matrix{Int}, Matrix{Float64}, Matrix{String}})
     posterior = list(le_nlv, Matrix{Float64})
     @inbounds for i = 1:le_nlv
         znlv = nlv[i]
